fix: localstorage, refactor
